test 1 + 2
The initial objective is to update the text value found within the component DynamicText.tsx by calling the function changeValue and passing a value to it.
The changeValue function have to be called from the index.tsx file inside the onChange which is called by the input component. The value of the input should match as the text displayed in DynamicText.tsx.
This is to be completed by only using React references. No props, context or external library should be used to solve this problem. Only React references should be used by referencing the DynamicText component and calling the changeValue within it.
Make the DynamicText component responsive to break if the word gets to long and goes beyond the width. Use the Chakra UI library.